computer_vision/pyproject.toml
[project]
name = "from-scratch-nn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=3.6.0",
    "evaluate>=0.4.3",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "pillow>=11.2.1",
    "pydantic>=2.11.4",
    "scikit-learn>=1.6.1",
    "tensorboardx>=2.6.2.2",
    "tqdm>=4.67.1",
    "transformers[torch]>=4.51.3",
]
